fix(common): refactor code structure
