+++ product-service/serverless.yml
@@ -46,5 +46,5 @@ custom:
     title: 'Product Service API'
     apiType: 'httpApi'
     generateSwaggerOnDeploy: true
-    typefiles: [ './src/types/products.d.ts' ]
+    typefiles: [ './src/types/api-types.d.ts' ]
     schemes: [ 'https' ]
