+++ product-service/src/common/http/response.ts
@@ -1,0 +1,9 @@
+export const DEFAULT_AWS_GATEWAY_API_RESPONSE_HEADERS = {
+  "Access-Control-Allow-Origin": "*",
+  "content-type": "application/json"
+}
+
+export enum RESPONSE_STATUS_CODES {
+  OK = 200,
+  NOT_FOUND = 404
+}
