+++ product-service/src/common/index.ts
@@ -1,0 +1,1 @@
+export { DEFAULT_AWS_GATEWAY_API_RESPONSE_HEADERS, RESPONSE_STATUS_CODES } from "./http/response"
