+++ product-service/src/functions/get-products-list.test.ts
@@ -1,6 +1,7 @@
 import { APIGatewayProxyEvent, Context } from "aws-lambda";
 import { getProductsList } from "./get-products-list";
 import PRODUCTS from "../mock/products.json";
+import { RESPONSE_STATUS_CODES } from '../common';
 
 describe('getProductsList function', function () {
   test('should return successful response', async () => {
@@ -14,7 +15,7 @@ describe('getProductsList function', function () {
 
     const result = await getProductsList(event, context)
 
-    expect(result.statusCode).toEqual(200);
+    expect(result.statusCode).toEqual(RESPONSE_STATUS_CODES.OK);
     expect(result.body).toEqual(`${JSON.stringify(products)}`);
   });
 });
