+++ product-service/src/types/api-types.d.ts
@@ -1,0 +1,14 @@
+export interface Error {
+  message: string;
+}
+
+export interface Product {
+  count: number;
+  description: string;
+  id: string;
+  price: number;
+  title: string;
+}
+
+export type Products = Array<Product>
+
