--- product-service/src/types/products.d.ts
@@ -1,13 +1,0 @@
-export type Product = {
-    count: number;
-    description: string;
-    id: string;
-    price: number;
-    title: string;
-}
-
-export type Products = Array<Product>
-
-export type Error = {
-    message: string;
-}
