+++ product-service/tsconfig.json
@@ -14,10 +14,10 @@
     "resolveJsonModule": true,
     "lib": ["es2020", "dom"],
     "rootDir": "./src",
-    "include" : ["src/types/**.d.ts"],
+    "include": ["./src/**/*.ts", "./src/types/api-types.d.ts"],
     "exclude": [
-      "src/**/*.spec.ts",
-      "src/**/*.test.ts"
+      "**/*.spec.ts",
+      "**/*.test.ts"
     ]
   }
 }
